Change git diff tooling
Remove delta. It wasn't that great and with some minor tweaks, the
default git diff view is a bit more readable. In order to make it more
readable, tweak the colours.

For when the default git diff view isn't good enough, add some difftool
config. Specifically neovim (via vimdiff) is the default difftool. Also
add configuration to use vscode as a GUI difftool.